YouTuberLoader/requirements.txt
fal-client
ipython
more-itertools
numpy
openai-whisper
opencc-python-reimplemented
optimum
pytubefix
python-dotenv
tiktoken
torch
transformers
tqdm